requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
bidict==0.21.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
colorama==0.4.4
h11==0.12.0
idna==2.10
multidict==5.1.0
pycodestyle==2.7.0
python-engineio==4.2.0
python-socketio==5.3.0
requests==2.25.1
toml==0.10.2
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
yarl==1.6.3
